Separate Client from Main
